Finishing of presentation Layer. Fix Repo. Fix DI. Change sorting option from Size to Rating

+++ app/src/main/java/duarte/pereira/appbasket/feature_basket/domain/model/AppItem.kt
@@ -5,7 +5,7 @@ data class AppItem(
     val packageName: String,
     val icon: String,
     val graphic: String,
-    val id: Int?,
+    val id: Long?,
     val downloads: Long,
     val rating: Double,
     val size: Long
